Move config of background jobs for rspec to jobs generator.
