+++ lib/suspenders.rb
@@ -1,7 +1,6 @@
 require "suspenders/version"
 require "suspenders/generators/app_generator"
 require "suspenders/generators/enforce_ssl_generator"
-require "suspenders/generators/initialize_active_job_generator"
 require "suspenders/generators/static_generator"
 require "suspenders/generators/stylesheet_base_generator"
 require "suspenders/generators/forms_generator"
@@ -9,6 +8,7 @@ require "suspenders/generators/ci_generator"
 require "suspenders/generators/db_optimizations_generator"
 require "suspenders/generators/factories_generator"
 require "suspenders/generators/lint_generator"
+require "suspenders/generators/jobs_generator"
 require "suspenders/actions"
 require "suspenders/adapters/heroku"
 require "suspenders/app_builder"
