+++ lib/suspenders/app_builder.rb
@@ -230,10 +230,6 @@ config.public_file_server.headers = {
       raise_on_missing_translations_in("test")
     end
 
-    def configure_background_jobs_for_rspec
-      run 'rails g delayed_job:active_record'
-    end
-
     def configure_action_mailer_in_specs
       copy_file 'action_mailer.rb', 'spec/support/action_mailer.rb'
     end
@@ -265,13 +261,6 @@ Rack::Timeout.timeout = (ENV["RACK_TIMEOUT"] || 10).to_i
         "production",
         %{ENV.fetch("ASSET_HOST", ENV.fetch("APPLICATION_HOST"))},
       )
-    end
-
-    def configure_active_job
-      configure_application_file(
-        "config.active_job.queue_adapter = :delayed_job"
-      )
-      configure_environment "test", "config.active_job.queue_adapter = :inline"
     end
 
     def generate_rspec
